+++ contracts/interfaces/pool/ICrocSwapPoolEvents.sol
@@ -86,5 +86,5 @@ interface ICrocSwapPoolEvents {
     /// @param recipient The address that receives the collected protocol fees
     /// @param amount0 The amount of token0 protocol fees that is withdrawn
     /// @param amount0 The amount of token1 protocol fees that is withdrawn
-    event CollectProtocol(address indexed sender, address indexed recipient, uint128 amount0, uint128 amount1);
+    event CollectProtocol(address indexed sender, address indexed recipient, uint256 amount0, uint256 amount1);
 }
